FlaskBlogHelper/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-wtf>=1.2.2",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-login>=0.6.3",
    "pymongo>=4.12.1",
    "markdown>=3.8",
    "pymongo-inmemory>=0.5.0",
    "django>=5.2.8",
    "pillow>=12.0.0",
]
